pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-apply"
version = "0.1.0"
description = "Automated job application tool"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40",
    "gspread>=6.0",
    "google-auth>=2.0",
    "playwright>=1.40",
    "pydantic>=2.5",
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.26",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.6",
    "pymupdf>=1.23",
    "reportlab>=4.0",
]

[project.scripts]
auto-apply = "src.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
